existsInTable integrated successfully into quotes model.

+++ api/categories/delete.php
@@ -5,18 +5,26 @@ $category = new Category($db);
 // Get raw posted data
 $data = json_decode(file_get_contents("php://input"));
 
-// Set ID to update
-$category->id = $data->id;
-
-// Delete post
-if($category->delete()) {
-  echo json_encode(
-  array(
-    'id' => $category->id
-  ));
-} else {
+// Check for required params
+if (isset($data->id)) {
+  // Set id to delete
+  $category->id = $data->id;
+  
+  // Delete author
+  if($category->delete()) {
+    echo json_encode(
+    array(
+      'id' => $category->id
+    ));
+  } else {
     echo json_encode(
     array(
       'message' => 'category_id Not Found'
     ));
+  }
+} else {
+  echo json_encode(
+  array(
+    'message' => 'Missing Required Parameters'
+  ));
 }
